Fix TypeError when using kwargs['cursorclass']
The cursorclass kwarg is automatically resolved from the
string name passed in the dict, however this causes a TypeError
when a connection is made more than once (that is, on two different
requests). This commit makes the routine check whether the argument
is a string, and in this way makes sure that the value has already
been resolved. This also allows for the user to pass a cursor class
directly.